libopkg/opkg_conf.c: Unitialized variable cleanups.
In OPKG 0.3.0, the '_conf' structure (referenced via opkg_config)
is created as a static variable, but never initialized.

During certain cleanup operations, the following pattern was used to
deallocate string memory:

    tmp_val = (char **)options[i].value;
    if (*tmp_val) {
        free(*tmp_val);
        *tmp_val = NULL;
    }

This works, but only if options[i].value is zero-valued by default.
On Linux, the memory segment it goes into is zero-filled prior to
program launch, so was a non-issue. On Cygwin, it causes a segfault.

This commit adds a zero-fill during opkg_conf_init(), and also ensures
that all string pointers are set back to NULL after they are free()'ed.

Signed-off-by: Nick Clark <nicholas.clark@gmail.com>