opkg/libopkg: Added config-file option to ignore UIDs
On filesystems where UIDs aren't well-behaved, UIDs/GIDs can't
always be trusted to be restorable. Additionally, non-privileged
users who want to use OPKG to install files in some home directory
won't generally be able to control the permissions of the output.

OPKG's default behaviour when faced with a UID mismatch is to report
the warning that libarchive produces, but to continue anyways.

This patch doesn't change the default behavior, but it adds an
'ignore_uid' boolean option to opkg_conf. If enabled in the config
file by 'option ignore_uid', libarchive will be called without the
'ARCHIVE_EXTRACT_OWNER' flag (and UID/GID restoration won't be
attempted).

Signed-off-by: Nick Clark <nicholas.clark@gmail.com>