Libopkg: Replaced malloc.h with stdlib.h.
This commit replaces all instances of #include <malloc.h> with
This commit fixes builds on OS X (under Macports, at least), and
is necessary because OS X doesn't provide malloc.h in the same
location as Linux. Stdlib, however, is compiler-provided, and is
the canonical home of malloc() and friends, as well as free().

Signed-off-by: Nick Clark <nicholas.clark@gmail.com>